Move lmnet/lmnet/common.py to blueoil/common.py
