+++ blueoil/common.py
