+++ lmnet/lmnet/datasets/camvid.py
@@ -20,7 +20,7 @@ import numpy as np
 import pandas as pd
 from PIL import Image
 
-from lmnet.common import get_color_map
+from blueoil.common import get_color_map
 from lmnet.datasets.base import SegmentationBase, StoragePathCustomizable
 from lmnet.utils.random import shuffle, train_test_split
