+++ lmnet/lmnet/utils/config.py
@@ -23,7 +23,7 @@ from tensorflow.io import gfile
 from yaml.representer import Representer
 
 from lmnet import environment
-from lmnet.common import Tasks
+from blueoil.common import Tasks
 from lmnet.data_processor import Processor, Sequence
 
 PARAMS_FOR_EXPORT = [
