+++ lmnet/lmnet/utils/predict_output/output.py
@@ -23,7 +23,7 @@ import numpy as np
 import PIL.Image
 import PIL.ImageDraw
 
-from lmnet.common import Tasks, get_color_map
+from blueoil.common import Tasks, get_color_map
 from lmnet.visualize import visualize_keypoint_detection
