+++ lmnet/tests/fixtures/configs/for_build_tfds_classification.py
@@ -16,7 +16,7 @@
 import tensorflow as tf
 from easydict import EasyDict
 
-from lmnet.common import Tasks
+from blueoil.common import Tasks
 from lmnet.datasets.delta_mark import ClassificationBase
 from lmnet.networks.classification.lmnet_v0 import LmnetV0Quantize
 from lmnet.data_processor import Sequence
