+++ lmnet/tests/fixtures/configs/for_build_tfds_object_detection.py
@@ -16,7 +16,7 @@
 import tensorflow as tf
 from easydict import EasyDict
 
-from lmnet.common import Tasks
+from blueoil.common import Tasks
 from lmnet.datasets.delta_mark import ObjectDetectionBase
 from lmnet.networks.object_detection.lm_fyolo import LMFYoloQuantize
 from lmnet.data_processor import Sequence
