+++ lmnet/tests/fixtures/configs/for_export.py
@@ -16,7 +16,7 @@
 from easydict import EasyDict
 import tensorflow as tf
 
-from lmnet.common import Tasks
+from blueoil.common import Tasks
 from lmnet.datasets.image_folder import ImageFolderBase
 from lmnet.networks.classification.lmnet_v0 import LmnetV0Quantize
 from lmnet.data_processor import Sequence
