+++ lmnet/tests/fixtures/configs/for_predict_object_detection.py
@@ -16,7 +16,7 @@
 from easydict import EasyDict
 import tensorflow as tf
 
-from lmnet.common import Tasks
+from blueoil.common import Tasks
 from lmnet.datasets.pascalvoc_2007 import Pascalvoc2007
 from lmnet.networks.object_detection.lm_fyolo import LMFYoloQuantize
 from lmnet.data_processor import Sequence
