+++ lmnet/tests/lmnet_tests/networks_tests/object_detection_tests/test_yolo_v2.py
@@ -20,7 +20,7 @@ from easydict import EasyDict
 
 from executor.train import start_training
 from lmnet import environment
-from lmnet.common import Tasks
+from blueoil.common import Tasks
 from lmnet.data_processor import Sequence
 from lmnet.datasets.pascalvoc_2007 import Pascalvoc2007
 from lmnet.networks.object_detection.yolo_v2 import YoloV2
