+++ lmnet/tests/lmnet_tests/networks_tests/object_detection_tests/test_yolo_v2_quantize.py
@@ -19,7 +19,7 @@ from easydict import EasyDict
 
 from executor.train import start_training
 from lmnet import environment
-from lmnet.common import Tasks
+from blueoil.common import Tasks
 from lmnet.datasets.pascalvoc_2007 import Pascalvoc2007
 from lmnet.networks.object_detection.yolo_v2_quantize import YoloV2Quantize
 from lmnet.pre_processor import ResizeWithGtBoxes
