+++ lmnet/tests/lmnet_tests/networks_tests/segmentation_tests/test_lm_bisenet.py
@@ -20,7 +20,7 @@ from easydict import EasyDict
 
 from executor.train import start_training
 from lmnet import environment
-from lmnet.common import Tasks
+from blueoil.common import Tasks
 from lmnet.data_processor import Sequence
 from lmnet.datasets.camvid import Camvid
 from lmnet.networks.segmentation.lm_bisenet import LMBiSeNet
