+++ lmnet/tests/lmnet_tests/util_tests/test_predict_output.py
@@ -20,7 +20,7 @@ from io import BytesIO
 import numpy as np
 import PIL.Image
 
-from lmnet.common import Tasks
+from blueoil.common import Tasks
 from lmnet.utils.predict_output.output import JsonOutput
